Made all menu screen extend abstractmenuscreen
